Add ref-counting for analysis units
Because of upcoming language bindings (in particular for languages with
automatic memory management), users might keep references to analysis
units after analysis context destruction. In such cases, we need to keep
them allocated not to crash.

This commits introduces reference counting for them so that bindings
have the necessary bits to implement this behavior.

Preparatory work for O323-002

Change-Id: Ie73527df575c3deea5f75b42a4e79c40996a28ae