Introduce wrappers for specific parsing primitives
Followup on O522-064

These wrappers replace the low-level ones that used to be exposed to
get closer to a nice user parsing API that takes care of cleaning memos,
increment the ref. count and synthetize diagnostics in case of parsing
error.

As a nice side effects, this fixes the exception we had on parsing
errors in the "parse" testing program.

Change-Id: Ic94020ce1d8603fc44ac2539848b0edf1e133222