Langkit: enhance doc for inherited fields
Field accessors are generated only for the AST node subclass that define
them (i.e. not for subclasses that inherit fields).  When using the
generated AST documentation in order to explore some tree, it is thus
useful to know which fields are inherited and from which class they
come. This change enhances this documentation to include this
information.

In order to implement this feature, this change also adds a new
"ast_node" attribute to langkit.compiled_types.Field instances that
indicates for each field what ASTNode subclass declared it.

Change-Id: I4944576c4b3f65dbb2b33e9a12690eea1c175237
TN: OA19-039