Langkit: make Unicode conversions out of Quex
... so that we can parse files bigger than 64KB. There seems to be some
bug in Quex that makes it believe that the source buffer ends at 64KB.
Performing Unicode decoding before running the lexer is a workaround
since Quex does not need to manage a conversion buffer anymore.

An additional benefit is that using directly GNATCOLL.Iconv makes it
possible to handle transcoding errors. This will land in an upcoming
commit.

TN: O414-018
Change-Id: I4e76bf0c5d8ce13c3ba96474dff16960aced99f7