Share documentations between API bindings
Before this change, the documentation for API entities was duplicated
across the codebase: in the original Ada API, in the C-compatible Ada
API, in the C header file and it would be duplicated in the Python
wrapper module if we had documentation there.

This commit introduces a new module to contain documentation for all
these entities along with comment/docstring formatting code. It also
updates the relevant templates to use this new facility.

Change-Id: I68565386aaf2576ffd3a83143aa1e849a13816b0
TN: OB06-010