Properties: add mapping and filtering expressions
This commit introduces new expression primitives in the properties DSL
that enable one to filter, transform and concatenate "collections"
values. What we call collections in this context is either an ASTList,
or an array type (which this commits introduces to compiled_types). The
C and Python API bindings are also updated to properly handle these
array types.

Change-Id: I871754508938b13f66c0db947790852e6a914cff
TN: OC10-027