Remove the base class for ASTList nodes.
This recent addition does not work well with the also recent removal of
lazy typing for structures and AST nodes: this base ASTList class was
considered as an ASTNode whereas it's not really (it’s a generic node,
so it cannot be typed!). Besides, it is not supposed to contribute to
code generation.

This base class was introduced to ease type checks in expressions, so we
just introduce a class attribute for that and a function to test whether
a CompiledType is a list simply.

Change-Id: Ifeedfecd85e41d2991806e103f0dedfa5507fb7e
TN: OA28-063