StructMetaClass: Refactor the root_grammar_class mechanism.
We now take the first class deriving from ASTNode as being the root
grammar class, and forbid further derivations directly from ASTNode. The
@root_grammar_class decorator just acts as an additional
check/documentation of the user code.

This will allow us to know in StructMetaClass when we are creating the
root grammar class, and to treat it in a special way (for example to add
built-in fields)

TN: OA28-063

Change-Id: Ib59474372aa38110f57e86a69647f07ed3407cdf