Expose parent links as properties
Before this change, the AST node parent link was hard-coded in the
C and Python APIs. This change moves it to the standardized properties
mechanism, so "parent" is now a built-in property for the root AST node.

It also adds a "parents" property to return the list of all parents of
some node.

Change-Id: I4d0fb8e43bf6ba8219e061ecdb862267faf451e5
TN: P106-015