Move the declarations of arrays/vectors of root AST node in the AST pkg.
An upcoming change will need the array type for root AST nodes in the
very same package as the root AST node is defined, so this change
introduce a kludge to make this possible while still keeping the
corresponding entities exported in the root generated package (with
subtypes and renamings).

This commit also turns the "array_type" function in compiled_types into
a CompiledType class method: it is handier and reduces the need for
modules to depend on compiled_types.

Change-Id: I465aec813cdb8b11fc0858ebaacbefd99d65ea52
TN: P106-015