Report unused parsing rules and unused AST node subtypes as errors
Unreferenced AST (actually untyped) nodes were already detected, so this
only enhances the error message to report them. On the other hand, the
unused parsing rules check is new. Also report parsing rules that are
defined multiple times.

This change was helpful for the process of updating the language
specification for GPR to cope with all changes in Langkit after the
repository split.

Change-Id: Ie4e21b03d08af7f43651663b58701dcd91d5db10
TN: P127-047