Introduce a Find AST primitive in Ada
This new primitive takes an AST node tree and a predicate and it returns
a lazy collection of all AST nodes under that tree that satisfy the
given predicate.

As it eased the implementation, it also introduces a Traverse
overloading that yields all AST nodes in a lazy fashion.

Change-Id: Ie2f07d08a7c04558ce25876e4b3111dde7bc517f
TN: P225-028