Properties: make "prepare", "freeze" and "constructs" passes separate
Properties accept lambda functions as expressions in the user DSL for
various purposes. These are later expanded into regular
AbstractExpression instances (this is the "prepare" pass). This
expansion requires the various nodes involved not to be frozen.

This is not possible right now since expansion for property A can happen
after the freezing of property B. As properties can share
AbtsractExpression instances, this does not work.

In order to make this scheme fully operational, we need to make it
possible to expand the lambda functions for all properties and *only*
then to freeze all AbstractExpression instances. This is what this
change does, introducing different passes for these steps.

Change-Id: Id00a4ee6bac103f480e23c32d1cce1e21562e298
TN: OA28-063