Analysis: make it possible to use alternate root parsing rules
... so that users do not need anymore to run a parser by hand in order
to parse, say, an expression. Also, this prepares the change that will
make all AST nodes require an analysis unit.

As a side effect, this change hides all parsers that do not return AST
nodes. This is required since analysis units can only hold AST nodes.
There is a good benefit for this: most of the generated code for the
"parser" interactive main could be removed.

Change-Id: Ic95b8f58620e45882670bd9db4b138db0a07bd15
TN: P308-028