Replace TDH in AST nodes with an analysis unit reference
As a consequence, add a way to get this TDH (token data handler) from
the analysis unit interface and then update all references to this TDH
in AST nodes primitives and in AST node parsers to go through the unit.

This opens the door of controlling the allocation/deallocation cycle of
lexical environments!

Change-Id: I13fcca4199946ac83b9db593665974294e163ecf
TN: P308-028