Remove multi-thread support for the C exception mechanism
The Thread_Local_Storage aspect is currently not available anymore on
Darwin platforms. Multi-threading support in Langkit is not tested, so
we cannot really say we support it.

This change temporarily removes this aspect so we build on Darwin again.
We will re-introduce it or use a task attribute instead when we will
decide to spend efforts to support multi-threading.

Change-Id: Ia620c5867beda83349355a70b9aa4857e18379de
TN: P317-027