Analysis_Interfaces: rename Deallocatable into Destroyable...
... and enhance the destroyable mechanism to expose a typed interface.

Before this change, this required clients to deal with System.Address
values and conversion wrappers in order to register objects.

This business has now been wrapped into a generic that
Analysis_Interfaces provides so that clients only have to instantiate it
and call the resulting procedure. As a result, the "untyped" interface
is now used in a very localized place while clients can deal with
properly typed access to objects and the corresponding destroy
procedures.

Change-Id: I069ddb1fb8475f43baf3ffe0755bcf439e89059d
TN: P308-027