Introduce some context for error diagnostics.
Expressions: Alter every abstract expression to call a new base
constructor that will extract the location of definition of the
expression.

AbstractNodeData: Do the same thing, so that we also have the
information of which Property/Field the error has happened in.

ASTNode.compute_properties: Introduce property context.

TN: P303-009

Change-Id: I700553f7084a41fa5519e650e2ca30577fbfb4aa