Represent tokens as indices in TDH tables instead of full records
This reduces the memory usage of token fields in AST node. They are
exposed as a tuple (analysis unit, token index) in the generated C API
and this tuple is wrapped conveniently in Python.

Due to code generation limitations (internal Ada types must be the same
when exposed in the public Ada API), it is not possible to have such a
convenient wrapping in the public Ada API for now. Hence, tokens are
exposed as token indices there as well. We should fix this one day.

Change-Id: Idca6d4f090c0ffdf24519b448e1040e1de4c124a
TN: P405-017