Allow c_api.CAPIType accept Name instances and update usages
So far, c_api.CAPIType's constructor has been accepting string names
only. However, Name instances were harmless when the .name() method was
called with a contextual default formatting.

Now, passing Name instances to this constructor is actually handy, so
allow it. This change adds the appropriate type annotations so that
this is not ambiguous anymore. This also simplifies multiple CAPIType
constructor calls to leverage this.

Change-Id: Ic3e12d13d49b9d5b8908cc8a3c6b3543471abf5a
TN: P405-017