By default, disable warnings when building the generated library
For new projects, a lot of entities are likely to be unreferenced in
the generated library. Enhancing code generation to avoid these cases
would be great, but it is likely to take a lot of efforts.

Instead, this commit introduces a couple of options in libmanage.py to
enable/disable these. They are disabled by default, but this can be
overriden in ManageScript subclasses.

Change-Id: I091b03a455d01a27502e39af6e0be68f857ac5da
TN: P412-008