Make it possible for properties to have different prefixes in code gen.
So far, all properties used to be compiled into functions called
"P_XXX". This change enhances the PropertyDef constructor to accept an
arbitrary prefix and updates the Property and AbstractProperty
constructors to keep the "P_" prefix.

We need this so that we can define automatically generated properties
whose names will not clash with user-defined ones.

Change-Id: If446892a41790a47ee5b29e6fcfeb968acf00f00
TN: P412-012