Add lexical env group operation in the properties DSL
This change adds a "group" operation in the properties DSL so that
properties can group together multiple lexical environment as a single
one. This operation is implemented as a function in
Langkit_Support.Lexical_Env which is merely bound with abstract
expressions.

This operation deals with arrays of lexical environments, so this change
also makes code generation for such arrays compulsory in the AST
package.

Change-Id: Iee912a41a6128194afe8adb7893a04650e62e85d
TN: P415-016