Turn Langkit_Support.Token_Data_Handler into a generic package
This makes it possible to merge Token_Raw_Data_Type and Token_Data_Type
into a single record type and to reduce the number of places where we
have to perform the conversion: mere integer -> proper token kind
enumeration value.

As a side effect, this also required to rename the Token_Data_Handler
unit into Token_Data_Handlers, so that there is no name conflict with
the Token_Data_Handler type it defines.

Change-Id: I55fae403573d331dab29b3a215be63975f44a25f
TN: P428-010