Properties: fix typing recursion detection
This fixes the properties/invalid_prop_type testcase, which failed
because of an infinite recursion after a recent change for this ticket
(I3851454019fe31275854bf1f3086ea9c965dda3e).

This change turned lazy type computing for field access resolved
expressions into eager computing. The latter was out of the scope of a
recursion guard... This fix extend the recursion guard to cover this
case.

Change-Id: I140f817d380d169955438b3ded31f25f1c21d82b
TN: OA28-063