Make it possible to yield exc. info. to APIs without raising fatal ones
Currently, the mechanism in the C/Python APIs to bind exception
information (exception message, sloc for the raise point) is available
only for the fatal errors, i.e. the ones that should not be ignored.

This change extends this mechanism to make it possible to bind exception
information also for non-fatal errors, such as Property_Error
exceptions.

Change-Id: Ibfec08261bff0108098a15c4bbe5b87cd4a44174
TN: P514-017