C API: make destroy_text accept a pointer to a text object
${text_type} is defined to have the C_Pass_By_Copy convention. The
destroy_text procedure takes an ${text_type} argument and has C
convention. However as this argument has IN OUT mode, it is actually
passed by reference, which is what the Python binding assumes. This is
confusing, so make it explicit in the Ada implementation that this
argument is actually a pointer.

Change-Id: I2766127df83c114b9eec36c1d0d7ce70c5a528d4
TN: P602-047