Move Bump_Ptr.Tagged_Alloc instances to $.AST.Types.Parsers' body
As we currently don't allow users to create/modify trees, exposing this
in a public API does not make sense, so move these instances to the only
place where they are actually used: parsers.

We were hopping that this would reduce the compilation time for users,
but this does not seem to change anything... We need profiling!

Change-Id: Icb98a336e79652dcf562a29a0110262263aa4955
TN: PB07-025