C/Python API: wrap Property_Error raised in Populate_Lexical_Env
Before this commit, these were considered as native exceptions (i.e.
Langkit bugs) whereas they clearly are bugs in the language
specification. Note that just like in the case of unit reparsing, we
need to invalidate lexical environments information when this happens:
this should be fixed in the future (there is a TODO for this).

Change-Id: I8fefe6fabe50f9c59ff7d8bbba1edb02cf8d4029
TN: PC02-013