Reset Env binding when constructing a property
As of today, the "construct" pass for some property can be triggered
in the middle of running this same pass on another property. When we do
this, information about the original property must not leak to the
compilation of the next property.

This principle was broken for Env binding: a property A could use the
Env binding that was not reset from the compilation of a property B.

Change-Id: Iceb7f426cedcb5eacdfb8da8691f2fbc5f2f7111
TN: PB18-030