Split single-node Destroy primitive and whole-tree one
As of today, the AST node destruction primitive (Destroy) destroys a
node *and* all its children recursively. For the upcoming node
synthetization feature, we would like to be able to destroy only a
single node at a time.

This change intoduces a new Destroy_Node primitive and make the current
Destroy one class-wide, to use Destroy_Node and to be the one doing the
recursion.

Change-Id: I1ba06a3c7e24c658340e6ed7082f9a41a26e09c9
TN: PB18-030