Introduce CompiledType.convert_to_storage_expr
Compiled types can be actually be mapped to two types in the generated
code: the default one and the one stored in AST nodes. For these, we
need to convert default-typed values to the storage-typed ones. This
commit intoduces a facility to do that.

This provides an implementation for tokens and leaves the one for logic
variables unimplemented.

Change-Id: If93240383dc2e6db4391ce338dfc2831c3116b14
TN: PB18-030