P506-012: Fix behavior of Bind with a conversion property
Before, the engine would try to use the conversion property regardless
of whether it was from L -> R or from R -> L. However, conversion
properties are generally valid only one way.

This implements a valid behavior for bind with a conversion property, as
an option in Adalog, and forces langkit to use this mode when the user
specifies a conversion property.

Change-Id: I16ff7519257d5ff086e0af520ff1bde514a0b5ec