Unparsers: fix classification of (un)optional tokens around fields
In parsers such as:

    Transform(Row(
        Pick('keyword1', Opt('keyword2', field_parser))
    ), NodeX)

The 'keyword1' token is mandatory, while 'keyword2' is present iff
field_parser manages to parse NodeX's only field. Because of this, in
unparsers generation, 'keyword1' must be a pre-token for the NodeX
unparser, rather than a pre-token for its field.

This commit fixes the parsers-to-unparsers transformation pass to
properly handle this case.

TN: R115-038