Ada public API refactoring
At a high level: $.Analysis is now (mostly) a wrapper around the
implementation. This means that there is be no dependency from the
implementation spec to the public API spec.

Practically: in order not to duplicate some stuff that does not need to be
duplicated, we allow the implementation ($.Implementation) AND the main
public API package ($.Analysis) to depend on shared declarations
($.Common), that are also part of the public API.

This creates the following (simplified) dependency graph:

    +------------+
    | $.Analysis | ---------\
    +------------+          |
           |                |    +----------+
           |                |--->| $.Common |
           v                |    +----------+
    +------------------+    |
    | $.Implementation |----/
    +------------------+

TN: R601-021