Remove the PLE queue
This rolls back an old change that meant to dissociate reparsing and the
update of corresponding lexical envs, so that these updates could be
done only when needed or on demand.

This behavior will conflict with analysis unit version-based safety
checks, as the analysis unit version is currently incremented in an
asynchronous way. So let's remove this mechanism.

It turns out anyway we don't really need this: that was for users to
chose when to trigger this (potentially costly) action, for instance to
avoid GUI loop pauses. We instead plan to advertise the use of LAL in a
thread for such use cases.

TN: R601-021