Turn Child_Index into a property
This matches more its nature (a function, not a field), which matters
for upcoming commits that add manual initialization for fields.

The only downside is that the return type of the corresponding function
in the public Ada API goes from Natural to Integer. We could probably
extend internals to allow this refinement in the future, but this is
very minor, so not blocking in the meantime.

TN: S311-015