S701-037: Fix invalidation of env_getter cache.
The Is_Stale procedure was missing information to determine if rebindings
from a given rebound_env where still valid or not and would unsoundly
consider them valid all the time.

This fix adds a field to rebound environments indicating the version of
the context at the time the rebound env is created. Is_Stale now compares
this number with the current context version to determine if a cached
rebound_env is stale or not.