Unparsing: keep preserving formatting in depth
This commit fixes the behavior of the unparsing regarding the
preservation of formatting when traversing the tree of rewritten nodes.
So far, processing a new node (i.e. one which was synthetized during
rewritting) disabled the preservation of formatting for all of its
children, even the ones that were not synthetized. This change
re-enables the preservation of such nested original nodes.

TN: S729-027