Move the Is_Env_Populated node field to the standard fields machinery
Since the move of bare nodes to untagged records, we lost the ability to
use default field initialization. The Is_Env_Populated field used to
rely on that, while now field initialization need to be manual in
initializer procedures. Converting this field from hardcoding in
templates to our field processing machinery allows us to automatically
initialize it in field init. procedures.

TN: S311-015