Use a single discriminated record type for all bare nodes
This commit changes the representation of bare nodes in the generated
code ($.Implementation). From distinct record types (one per node type),
we switch to a single discriminated record (discriminant is the kind).

In order to preserve separate type names for each node (including
abstract ones), this replaces the previous incompatible access types
with a single access type and one subtype for each node. A dynamic
predicates on all these subtypes allow us to rely on Ada semantics to
check at runtime (iff assertions are enabled) that expected node type
are used (for instance as actuals in function calls).

TN: S311-015