AbstractNodeData: introduce the api_name property
The upcoming unification of types for bare nodes requires to have
distinct names for internal property overloads and field accessors,
however we want to preserve original names in public APIs. Introduce
api_name to generate public APIs, so that we can modify the "name" to be
used in the rest of code generation.

TN: S311-015