AbstractNodeData: introduce the "indexing_name" property
In properties lowering, we end up with several properties which must
have the same original name: the dispatcher needs to preserve the
original name for debugging and the root static property must preserve
the original name for error messages. However we need both to appear in
the same nodes' "_fields" dict, so we cannot use "original_name" as the
key for fields in this dict.

This change solves this problem introducing (yet) another name:
"indexing_name", whose only purpose is to provide a name to use as a key
in fields dicts.

TN: S311-015