project_file.mako: fix handling of --no-gdb-hook
Until this commit, using this option would make the project_file.mako
template generate invalid GPR syntax: a trailing comma in the aggregate
for the Interfaces project attribute.

In order to fix this, we now compute the list of interfaces in
emitter.py, and just do formatting in the template.

TN: R411-040