Split $.Converters
In order to disable the generation of the public Ada API, we need to
separate the Converters unit in two parts: one that deals with public
API types (public converters), and one that doesn't (private
converters). All units that depend on the private converters will still
compile when the public Ada API is disabled.

TN: S823-010