$.Implementation: default to the analysis unit's charset for reparsing
When we reparse but don't ask for a specific charset, we should use the
charset that was used to parse the analysis unit last time. This commit
adjusts parsing code to work this way: we used not to, which made it
impossible to perform rewriting on non-ASCII sources.

TN: S903-011