Introduce ref-counting in internal unit providers
Internal unit providers (or rather accesses to them) are the entities
used at the C API boundary and in $.Implementation. As a consequence,
their lifetimes depend both on their use in C API users and both by the
lifetime of the analysis contexts that reference them.

In order to properly handle their resources allocation, this commit
introduces ref-counting primitives for internal unit providers so that
they are deallocated only when it is safe to do so.

TN: S910-015