S924-008 Fix unwrapping of enum types
An EnumType has a ctype view, thus, to unwrap it, it is unecessary to
call the unwrap function.

The wrap function was not being affected by the issue.

wrap and unwrap functions are required only if a type does not define a
ctype view. Thus, refactor the code to reflect this.