Rework the layout of obj/lib build subdirectories
This turns the build directory layout for object and library files from:

  * lib/$NAME.$LIBRARY_TYPE/
  * obj/$NAME.$LIBRARY_TYPE/

to:

  * lib/$NAME/$LIBRARY_TYPE/$BUILD_MODE
  * obj/$NAME/$LIBRARY_TYPE/$BUILD_MODE

The new layout allows one to have dev and prod builds coexisting in the
same build directory. As a side effect, libmanage's setenv and install
commands now accept the --build-mode option, so that they can work with
the build artifacts from the intended build mode.

TN: SA30-026