Always build with -fPIC and share object directories
This patch changes all generated project files so that builds for all
library kinds share the same object directories, and thus every unit is
compiled only once (not once for static builds, then once for
relocatable builds, etc.). Always building with -fPIC allows us to do
this as -fPIC is a requirement for relocatable and static-pic builds.

Most generated library users use -fPIC-based builds, so the (limited?)
extra bloat in generated code for static libraries seems to be worth the
reduction of build times.

TN: S429-004