Recompute memoization key on memoization table reset
Memoization table invalidation can happen in the middle of the execution
of a memoized property, for instance when an external property loads a
new analysis unit.

Because of this, memoized properties must check cache version when
re-using memoization keys and cursors: on cache version bump,
memoization keys and cursors must be recomputed.

TN: SC09-060