Consider lexer state machine code as being part of the lib interface
For performance reasons, we do not regenerate $.Lexer_State_Machine when
the state machine has not changed (generating this huge file is slow).
This plays badly with the generation of the Interfaces project
attribute, as we currently "forget" to add $.Lexer_State_Machine to that
list when it is not regenerated.

This commit fixes that, so that this module is always part of the
Interfaces list, even when its code generation is skipped.

This bug was harmless so far, but gprbuild warns about this
inconsistency with the incoming code coverage setup.

TN: SC19-030