Coverage: create a dummy project to store gnatcov input sources
This will make it possible to make coverage computation (i.e. when
running the testsuite) independent of the original library source tree.
This is required in order to compute a coverage report in our production
system.

As a bonus, this simplifies the interface for testsuite scripts (no more
obscure --gnatcov-args argument).

TN: SC19-030