RA22-015: Do a `check_semantic` pass directly in liblktlang
As part of this, a new `LKNode.check_semantic` property is introduced.
This property recursively explores the whole subtree to run name and
type resolution on it. If called on the unit root, it will run
resolution for the whole unit.

This property returns a TreeSemanticResult, that contains both a list of
results (type and name resolutions, and errors), and a boolean flag
indicating if the subtree contains any errors.

`LangkitRoot.check_legality` is also introduced. This will run semantic
checks for all nodes and return only the diagnostics.