Keep foreign nodes sorted in lexical envs
The order of results in lexical env lookups is significant, so it is not
acceptable to have the order in which units were load have an influence
on the order of items in lexical environments.

This commit introduces a sorting mechanism for foreign nodes in lexical
envs to fix that.

TN: T320-010