C API: fix signature of array method arguments in header
Note that we never noticed them because almost no testcase actually
exercizes the C API through the C header: either tests use the Ada API
or the Python API (where the C header is shortcircuited). It is not
clear at the moment whether keeping this C header is useful.

TN: T423-042