lkt_lowering: switch to a more argparse-like API
Turn the result of parsing annotation from a dict to a namespace object.
In addition, allow default values for annotations. This allows simple
idioms to get an actual boolean from a FlagAnnotationSpec even when the
annotation is absent. The following:

    annotations.get("annot_name")

will return either False (when absent) or True (when present).

TN: RA22-015