Lkt lowering: replace Environment with a simple dict
The only purpose of the Environment dict was to handle scopes for local
entities in properties. This is not necessary as Lkt is supposed to
resolve locals references to their declaration. Just use a simple dict
instead. Note that we keep a separate dict for variables to keep types
homogeneous in each dict.

TN: RA22-015