Python API: simplify the binding of diagnostics
The lazy decoding of diagnostics that DiagnosticsList implements has no
tangible advantage right now, and its current form gets in the way of
typing the binding, as DiagnosticsList implements partially the
iteration protocol. Switch to eager decoding and a real list type
instead. As a bonus, this simplifies the binding.

TN: T619-037