SA28-035: Move Text_Buffer.Get_Line to Token_Data_Handlers
Also create a Text_Buffer_Ifc base class that is used in
Langkit_Support.Diagnostics.Output instead of the Text_Buffer type.

While this is better design than the previous design, because we can now
get rid of the AnalysisUnit.Text that returns an unsafe access, the end
goal is to be able to reuse the vector to calculate slocs, which will
allow us to remove slocs from tokens, only keeping the offset.