Run pre-initial env actions only during PLE
The goal of these actions (consecutive Do instances present first in
action list) is to fetch other units, so running them only once is
enough (during PLE, rather that each time the initial env is computed).
This will also restore simple semantics, as a preparation to the
introduction of named environments, which will slightly change the
behavior of SetInitialEnv actions.

TN: T320-010