SA28-035: Degenericize LK.Symbols/LK.Token_Data_Handlers
This is necessary if we want to have diagnostics generated during
lexing/parsing, that don't store slocs but offsets, and so need to have
a reference to the token_data_handler.

So far this type was generic, so we couldn't store a ref to it except by
making Langkit_Support.Diagnostics generic too, which we don't want to.