$.Implementation: always generate env spec helpers
Put the declaration of these helper subprograms in $.Implementation's
spec so that we never get a warning about them being unused. This is now
the most common way to handle templating: first it removes the need to
do complex conditional code emission, simplifying templates, and second,
in the future, these helpers will belong to Langkit_Support: avoiding to
generate them when they are not needed (i.e. only for toy language
specs) is pointless.

TN: T320-010