Use gnatcov's --no-subprojects flag
We want coverage analysis to process only the root project (i.e. the
generated library, and not langkit_support & all). This used to be the
default, but gnatcov recently switched to a "recursive" mode by default.
Use the new --no-subprojects flag to get back to the previous behavior.

TN: S528-004