Lexical envs: refactor to prepare for dynamic primary envs
An upcoming change will introduce a new kind of primary lexical
environment, which contains no "static" map but instead a reference to a
property that returns env associations (hence "dynamic").

This commit renames the "Primary" kind to "Static_Primary" and does
minor refactorings to prepare the ground for dynamic primary envs.

TN: T319-033