Stop depending on the funcy non-standard module
Each dependency outside the standard library slightly increases the
maintenance cost and the difficulty for new contributors.

Readability is always subjective, but performances are critical for
langkit and the changes introduced by this commit should sligthly
improve it:
* Several intermediate list constructions are replaced by integer
  list/tuple indices or Python generators.
* In langkit/expressions/base.py, traversal of a tree by a recursive
  generator is replaced with a single loop.  Moreover, according to
  https://github.com/Suor/funcy/blob/master/funcy/tree.py, the funcy
  implementation uses a deque structure while only right-appends are
  actually needed.

The only dubious change in langkit/dsl_unparse.py constructs an
intermediate tuple, but the previous version was concatenating
intermediate lists, which is slightly worst.