Fix wrong assumption in codegen for node field initializers
The computation of fields from the parent node assumed so far that in
the list of fields, parent fields came first and only then fields for
the current node. This assumption is wrong, because user fields can be
added to the parent's list of fields after the current node is assigned
its own fields.

This commit removes this assumption, switching from list indexes to set
operations in order to compute parent fields.

TN: T319-033